Make sure bulk import works for empty databases
We always look for the latest timestamp for a given node in the
database to make sure that we only import newer data. If there
was no data for the given node in the db yet, the code would not
find anything and crash. We therefore now also handle this case
and don't filter anything if nothing is in the db yet